docs/requirements.txt

# -- --------------------------------------------------------------------------------------------------- -- #
# -- MAF1731 - MICROSTRUCTURE AND TRADING SYSTEMS                                                        -- #
# -- requirements.txt: A text file with python libraries required for the project                        -- #
# -- Author: IFFranciscoME - if.francisco.me@gmail.com                                                   -- #
# -- License: MIT License                                                                                -- #
# -- Repository: https://github.com/iffranciscome/myst/                                                  -- #
# -- --------------------------------------------------------------------------------------------------- -- #

# Run the following in bash console:
# $ pip instal -r requirements.txt

# -- Generic
pandas>=1.1.4
numpy>=1.20

# -- Object Description
rich>=9.5

# -- Visualization
jupyter>=1.0
matplotlib>=3.4.1
seaborn>=0.11.1
plotly>=4.14

# -- Documentation
jupyter-sphinx==0.3.2
sphinx-rtd-theme==0.5.2
